pull form config objs into formConfigs file
